Add files via upload
This renames extensions. It ask what extension is to be changed, and what it's to be changed to. The option is given to keep the original files or delete them. This must be run within the same folder that the affected files are in as there is no option for folder specification.