Update README.md
Renames extensions in batches in a single folder.
This renames extensions. It ask what extension is to be changed, and what it's to be changed to. The option is given to keep the original files or delete them.

The 2 versions are as follows:

rename_extensions_1.1 - This must be run within the same folder that the affected files are in as there is no option for folder specification. This is for ease of use.

rename_extensions_dir_ask - The directory with the affected files must be specified upon prompt.